requirements.txt
always have 2 different terminals for backend and frontend

venv for django (django REST):
- cd advisuri (folder)
- python3 -m venv venv (creating virtual environment)
- env\Scripts\activate 

and then install dependencies in the virtual environment (prefix before the terminal line):
- (pip install -r dependencies.txt)
asgiref 
django 
django-cors-headers
djangorestframework 
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
axios
jwt-decode
react-router-dom

backend google auth (pip install):
django-allauth
google-auth
requests

frontend google auth:
npm install @react-oauth/google

#to start project (in env)
- django-admin startproject (backend name)
- cd (backend name)
- python manage.py startapp api (manage api)

# always activate the backend when running the server
to run the server:
- env\Scripts\activate
- python manage.py runserver


 #                        After the setup is done
#Frontend
- cd frontend
- npm run dev

#Backend
- cd backend
- python manage.py runserver


For typescripts:
npm install --save-dev @types/react @types/react-dom @types/react-router-dom
npm install --save-dev typescript-eslint
npm install clsx tailwind-merge #library needed in the UI section
npm install @radix-ui/react-accordion lucide-react
npm install @tailwindcss/vite tailwindcss
npm install tw-animate-css

Restart the Editor for Typescript

Press Ctrl + Shift + P (or Cmd + Shift + P on Mac).
Type "Restart TS Server" and press Enter.